Done with simple pirate. I am satisfied with this. Unlike in the book I use a dealer socket and do two phases of waiting. The way timeouts work turns out to be surprising. Even without simulations, when there are 3x more clients, having a timeout/2.0 time to do the work will have them all abort in the end.
You'd expect that as clients drop off, the rest would squeeze in, but the network congestion kills everyones it seems. Probably the workers get into a state when they are always working on stale requests. One thing that might have been worth a try is to send the time along with the message and dispose of the stale ones. But that is getting to fancy with these toy examples.